Conflicts:
	src/main/java/com/revature/kyubbs/models/Post.java
	src/main/java/com/revature/kyubbs/models/Report.java
	target/classes/com/revature/kyubbs/models/BoardThread.class
	target/classes/com/revature/kyubbs/models/Post.class
	target/classes/com/revature/kyubbs/models/Report.class
	target/classes/com/revature/kyubbs/services/PostServiceImpl.class

+++ src/main/java/com/revature/kyubbs/repositories/ReportRepository.java
@@ -8,8 +8,8 @@ public interface ReportRepository extends JpaRepository<Report, Long> {
 
 	public Report findReportByIp(String ip);
 
-	public Report findReportByPostId(int id);
+	public Report findReportByPostId(Long id);
 
-	public Report findReportByThreadId(int id);
+	public Report findReportByThreadId(Long id);
 
 }
